18/01/2009 (1.8.3.283) ---------- - Assets: ItemInfo.Volume changed from integer to double - Assets: Added a refining mode combobox - Assets: Added code to use overrides and their values - Assets: Corp Offices now taken into account - Assets: Added TabControl for splitting results - Assets: Total recycling results now displayed - Assets: Taken into account skills for wastage and for station take

+++ EveHQ.Assets/PlugInData.vb
@@ -96,7 +96,7 @@ Public Class PlugInData
                             newItem.MarketGroup = 0
                         End If
                         newItem.Published = CInt(itemRow.Item("published"))
-                        newItem.Volume = CInt(itemRow.Item("volume"))
+                        newItem.Volume = CDbl(itemRow.Item("volume"))
                         newItem.PortionSize = CInt(itemRow.Item("portionSize"))
                         Items.Add(newItem.ID.ToString, newItem)
                     Next
